Only access sys.user data when the '--owner' option is passed on CLI
This is a correction of the previous commit in which sys.user was always joined
to sys.table and sys.schema. The data from sys.user is used to filter queries by
owner. It is not necessary to perform a join with sys.user unless the filter
option '--owner' is added.